requirements.txt
bokeh==1.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.2
oauthlib==3.0.1
packaging==19.0
pandas==0.24.2
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-twitter==3.5
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tornado==6.0.2
urllib3==1.24.2
Werkzeug==0.15.2
wordcloud==1.5.0
